src/test/java/ru/job4j/concurrent/сompletableFuture/RolColSumTest.java
package ru.job4j.concurrent.сompletableFuture;


import org.junit.Test;

import java.util.concurrent.ExecutionException;

import static org.junit.Assert.*;
import static ru.job4j.concurrent.сompletableFuture.RolColSum.asyncSum;
import static ru.job4j.concurrent.сompletableFuture.RolColSum.sum;

public class RolColSumTest {
    private final int[][] res = {{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15},
            {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15},
            {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15},
            {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15},
            {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15},
            {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15},
            {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15},
            {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15},
            {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15},
            {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15},
            {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15},
            {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15},
            {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15},
            {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15},
            {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15}};

    @Test
    public void whenSumSecondLineConsistent() {
        RolColSum.Sums[] sums = sum(res);
        int result = sums[1].getRowSum();
        assertEquals(120, result);
    }

    @Test
    public void whenSumFirstColumnAsynchronous() throws ExecutionException, InterruptedException {
        RolColSum.Sums[] sums = asyncSum(res);
        int result = sums[0].getColSum();
        assertEquals(15, result);
    }
}